Rename _aux to _sysfonts
